Add files via upload
The Josephus Problem is also a theoretical problem studied by computer scientists and mathematicians. In this version, the people are standing in a circle and they are numbered, starting at 1 at the “head” of the circle, and ending with n, the number of people. Then, two numbers are picked out of a hat. The first number represents the person to start with. The second number represents the number to count before stopping on the person to be killed. After the person is killed, their body is moved outside the circle and is no longer counted. The counting and killing continue until there is only one person left.