Add files via upload
Basin of attraction algorithm. A basin of attraction for a root is the set of x values that converges to that root under Newton iterations. In this project, I produced colored plots showing the basins of attraction for the following functions. f(x) = (x - 4) * (x + 1), g(x) = (x - 1) * (x + 3),  h(x) = (x - 4) * (x - 1) * (x + 3).