Add files via upload
Exploring the solution to the restricted two-body problem. In this problem, one of the masses is much larger than the mass of the other, for example, the Sun Jupiter system. Lagrange published a paper solving the restricted two-body problem. What he observed was that there were five points in the orbital plane of the two bodies that had quasi-gravitational stability. Those points are now called the Lagrange points or Lagrangian points.